+++ src/web/assets/cp/src/js/BaseElementSelectInput.js
@@ -491,14 +491,9 @@ Craft.BaseElementSelectInput = Garnish.Base.extend(
             await Craft.appendHeadHtml(data.headHtml);
             await Craft.appendBodyHtml(data.bodyHtml);
           })
-          .catch(({response}) => {
-            if (response && response.data && response.data.message) {
-              Craft.cp.displayError(response.data.message);
-            } else {
-              Craft.cp.displayError();
-            }
-
-            reject(response.data.message);
+          .catch((e) => {
+            Craft.cp.displayError(e?.response?.data?.message);
+            reject(e?.response?.data?.message);
           })
           .finally(() => {
             this.hideSpinner();